requirements.txt
django
djangorestframework
djangorestframework-simplejwt
requests
django-cors-headers